Use the new "UpdateNodeOperands" method to simplify LegalizeDAG and make it faster.  This cuts about 120 lines of code out of the legalizer (mostly code checking to see if operands have changed).
It also fixes an ugly performance issue, where the legalizer cloned the entire
graph after any change.  Now the "UpdateNodeOperands" method gives it a chance
to reuse nodes if the operands of a node change but not its opcode or valuetypes.

This speeds up instruction selection time on kimwitu++ by about 8.2% with a
release build.