The change to sidebar.php makes the website more efficient by asking the database for only what it needs instead of all the other junk that is stored in the night table.
The changes to NightlyTester.php, ProgramResults.php, fulltest.php, and
test.php reflect the new way we are storing dejagnu pass/fail/xfail results
in the datase. Rest assured, there is a hack in ProgramResults.php to
ensure we can still handle the old way these were stored in the database,
although it should be noted that this old way is missing signifigant data.
This "functionality" of losing data is the reason I redesigned the database.

The changes to individualgraph.php are to bring it into compliance with the
LLVM coding standard (mainly it converts tabs to two spaces).