Pull r53503 into Gaz:
Fix PR2536: a nasty spiller bug. If a two-address instruction uses a register
but the use portion of its live range is not part of its liveinterval, it must
be defined by an implicit_def. In that case, do not spill the use. e.g.

8   %reg1024<def> = IMPLICIT_DEF
12  %reg1024<def> = INSERT_SUBREG %reg1024<kill>, %reg1025, 2

The live range [12, 14) are not part of the r1024 live interval since it's
defined by an implicit def. It will not conflicts with live interval of r1025.
Now suppose both registers are spilled, you can easily see a situation where
both registers are reloaded before the INSERT_SUBREG and both target registers
that would overlap.