Treat exit statuses that indicate failure to execute the test program and exit statuses that indicate that the test program was terminated by a signal as failures.
Also, report failures by writing a string to the output file which
will trigger a diff when it is compared with output files from other
runs of the test.

And, having reported any detected errors in a way that will be
visible to the test system, always exit "successfully". RunSafely.sh's
own exit status is not used to determine if a test has passed or failed.