Pull r64124 into Dib:
Turns out AnalyzeBranch can modify the mbb being analyzed. This is a nasty
suprise to some callers, e.g. register coalescer. For now, add an parameter that
tells AnalyzeBranch whether it's safe to modify the mbb. A better solution is
out there, but I don't have time to deal with it right now.