--- Merging (from foreign repository) r65985 into '.': A    test/CodeGen/X86/2009-03-03-BTHang.ll U    lib/CodeGen/SelectionDAG/TargetLowering.cpp
The DAG combiner was performing a BT combine. The BT combine had a value of -1,
so it changed it into a 31 via the TLO.ShrinkDemandedConstant() call. Then it
would go through the DAG combiner again. This time it had a value of 31, which
was turned into a -1 by TLI.SimplifyDemandedBits(). This would ping pong
forever.

Teach the TLO.ShrinkDemandedConstant() call not to lower a value if the demanded
value is an XOR of all ones.