--- Merging (from foreign repository) r69795 into '.': U    gcc/llvm-convert.cpp
Chris's new record constructor code broke the x86-32
Ada build because it doesn't handle the case when the
size of a field is less than the size of the type of
the field.  Fix by truncating the initial value in this
case.  It may be that this is only valid if the discarded
bits are zero, but I'm not sure so I didn't add a check
for this.