--- Merging r77994 into 'gcc/objc/objc-act.c': U    gcc/objc/objc-act.c
Wow! This is not good. So, it looks like we needed to generate "i8* null" for
the last argument of an llvm.eh.selector call for the @finally block. Daniel
explained some of the horrid behavior of the exception system. After I recovered
from that nightmare, we determined that ObjC needed to handle the @finally as a
C++ cleanup method. Hence, the "i8* null". This is only needed in 64-bit mode.