Add a stop gap to Sema::CorrectTypo() to correct only up to 20 typos. This is to address a serious performance problem observed when running 'clang -fsyntax-only' on really broken source files.  In one case, repeatedly calling CorrectTypo() caused one source file to be rejected after 2 minutes instead of 1 second.
This patch causes typo correction to take neglible time on that file
while still providing correction results for the first 20 cases.  I
felt this was a reasonable number for moderately broken source files.

I don't claim this is the best solution.  Comments welcome.  It is
necessary for us to address this issue because it is a serious
performance problem.