disable a minor optimization on darwin.  For code like this:
extern int a, b, c, d;
void foo() {
  int *Arr[] = {&a, &b, &c, &d, &a, &b, &c, &d, &a, &b, &c, &d, &a, &b, &c, &d,
    &a, &b, &c, &d, &a, &b, &c, &d, &a, &b, &c, &d, &a, &b, &c, &d };
  bar(Arr);
}

the compiler synthesizes a global variable for the temporary array and then
memcpy's from it to Arr.  This is all great, but emitting the temporary global
with an L label on darwin fouls up atomization and can thus break dead code
stripping.

The right answer is to use an 'l' label without .globl, but this will require
introducing a yet-another new linkage type to llvm, which I'll do later.

This fixes rdar://7233622