Fix for PR 4262 / 6599374 (extern inline handling).
The basic problem with Duncan's previous attempt (see PR 4262)
is that there isn't anything in GCC's IR corresponding to "extern"
in the source; there is a DECL_EXTERNAL bit which means that in C,
but is abused by C++ and ObjC to mean other things.  The way around
that is to pay attention to that bit only if the language is C; ugly,
but so is the problem it's working around.  Tested by gcc.dg/98031[23]-1.c
in gcc testsuite.